+++ R/selection_criteria.R
@@ -120,6 +120,19 @@ Trule <- function(which_not_answered, possible_patterns, person, test, row_loc, 
     crit
 }
 
+Arule <- function(which_not_answered, possible_patterns, person, test, row_loc, method, design){
+    infos <- lapply(which_not_answered, function(x)
+        FI(extract.item(test$mirt_object, x), Theta=person$thetas))
+    acovs <- lapply(infos, function(x, person){
+        ret <- try(solve(x + person$info_thetas), TRUE)
+        if(is(ret, 'try-error')) ret <- diag(ncol(x)) * 10
+        ret
+    }, person=person)
+    crit <- lapply(acovs, function(x, w) sum(diag(x) * w), w=design$Wrule_weights)
+    crit <- do.call(c, crit)
+    crit
+}
+
 Wrule <- function(which_not_answered, possible_patterns, person, test, row_loc, method, design){
     infos <- lapply(which_not_answered, function(x)
         FI(extract.item(test$mirt_object, x), Theta=person$thetas))